backend/requirements.txt
#pip install 

dotEnv [ secure env ] 
google-auth
- may need to migrate to Gunicorn + unicorn worker or Daphne depends on how the app grows
pydantic - data validation library